Fixing up some more economy problems

+++ Content.Server/_CombatRim/Economy/Systems/SellPlatformSystem.cs
@@ -1,18 +1,10 @@
-
-using System.Collections.Generic;
-using System.Linq;
 using Content.Server.Cargo.Systems;
 using Content.Server.MachineLinking.Events;
 using Content.Server.Stack;
-using Content.Shared._Afterlight.Generator;
 using Content.Shared.Examine;
-using Content.Shared.Maps;
 using Content.Shared.Tag;
 using Robust.Server.GameObjects;
-using Robust.Shared.GameObjects;
 using Robust.Shared.Map;
-using Robust.Shared.Physics.Events;
-using Robust.Shared.Physics.Systems;
 
 namespace Content.Server._CombatRim.Economy
 {
@@ -42,27 +34,46 @@ namespace Content.Server._CombatRim.Economy
                 if (!TryComp<TransformComponent>(uid, out var xform))
                     return;
 
-                foreach (var otherUid in _entityLookup.GetEntitiesIntersecting(uid, LookupFlags.Dynamic))
+                var ecoComp = _economySystem.GetEconomyComponent();
+
+                if (ecoComp == null)
+                    return;
+
+                // No Static Flag, as we don't want players to just start selling walls. Sundries because otherwise you can't sell items on grids... but in space for some reason?
+                var intersecting = _entityLookup.GetEntitiesIntersecting(uid, LookupFlags.Sundries | LookupFlags.Dynamic);
+
+                var cash = 0;
+
+                foreach (var otherUid in intersecting)
                 {
                     if (otherUid == uid || (comp.Blacklist != null && comp.Blacklist.IsValid(otherUid)))
                         continue;
 
-                    var cost = _pricingSystem.GetPrice(otherUid)*comp.TaxCut;
+                    var cost = _pricingSystem.GetPrice(otherUid)*ecoComp.inflationMultiplier*comp.TaxCut;
 
                     if (!_economySystem.TryCapitalism((int) cost))
                         continue;
 
-                    var spawned = Spawn(comp.MoneyPrototype, xform.Coordinates);
-                    _stackSystem.SetCount(spawned, (int) cost);
-                    comp.Contacts.Remove(otherUid);
+                    cash += (int) cost;
                     _entityManager.DeleteEntity(otherUid);
                 }
+
+                if (cash <= 0)
+                    return;
+
+                var spawned = Spawn(comp.MoneyPrototype, xform.Coordinates);
+                _stackSystem.SetCount(spawned, cash);
             }
         }
 
         private void OnInspect(EntityUid uid, SellPlatformComponent comp, ExaminedEvent args)
         {
-            args.PushText(("Credits: " + _economySystem.credits));
+            var ecoComp = _economySystem.GetEconomyComponent();
+
+            if (ecoComp == null)
+                return;
+
+            args.PushText(("Credits: " + ecoComp.credits));
         }
     }
 }
